travis: keep es6 cache

+++ base/management/commands/init.py
@@ -31,6 +31,13 @@ class Command(BaseCommand):
             default=False,
             help='Do not attempt to compress static files.',
         )
+        parser.add_argument(
+            '--keep-cache',
+            action='store_true',
+            dest='keep-cache',
+            default=False,
+            help='Do not remove ES6 transpile cache dir if present.',
+        )
 
     def handle(self, *args, **options):
         if options["restart"]:
@@ -44,9 +51,10 @@ class Command(BaseCommand):
             "loaddata",
             "document/fixtures/initial_export_templates.json")
         call_command("compilemessages")
-        # Remove the es6 cache if it exists
-        if os.path.exists(os.path.join(PROJECT_PATH, "es6-cache")):
-            shutil.rmtree("es6-cache")
+        if not options["keep-cache"]:
+            # Remove the es6 cache if it exists
+            if os.path.exists(os.path.join(PROJECT_PATH, "es6-cache")):
+                shutil.rmtree("es6-cache")
         call_command("transpile")
         if not options["no-compress"]:
             try:
